backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
requests==2.31.0
google-generativeai>=0.7.2
python-multipart==0.0.9
python-json-logger==2.0.7

